feat(core): preserve disabled HTTP headers and params on dispatch
